boost.sh: use xcrun to locate the compiler, and the lipo and ar command line tools
the original script uses hardcoded paths that make assumptions about the inner
structure of an Xcode.app bundle. this is sure to fail as soon as Apple changes
something inside Xcode (there are reports that Xcode 5 already brings such a
change). a symptom of the problem is an error such as this:

sh: /Volumes/Xcode/Xcode.app/Contents/Developer/Toolchains/XcodeDefault.xctoolchain/usr/bin/clang++: No such file or directory