boost.sh: fix "toolset" parameter on bjam command line
Boost 1.54.0 has some internal changes that cause the simulator build to fail with
the current bjam command line that only uses "toolset=darwin". the solution is to
use different values for the "toolset=" parameter for the iPhone and the simulator
build. the value that we use now selects the appropriate configuration from the
user-config.jam file ("toolset=darwin-$IPHONEOS_BJAM_TOOLSET" and
"toolset=darwin-$IPHONE_SIMULATOR_BJAM_TOOLSET")

a symptom of the problem is when the compiler executable is not found and
the build fails with this error:

/bin/sh: line 1: -ftemplate-depth-128: command not found

the source of the fix is this Stack Overflow post:
http://stackoverflow.com/questions/17714635/how-can-i-compile-boost-1-54-0-1-54-for-the-ios-simulator-6-1-on-os-x-10-8-4