boost.sh: handle object files with same name in different libraries
it is possible that object files with the same name are present in multiple
libraries. for instance, utf8_codecvt_facet.o is present both in filesystem
and in program_options. the current decomposing mechanism cannot
handle this and clobbers the first object file with the second file. as a result,
the build of an iOS app that ends up consuming the boost framework
may fail with a weird linker error that complains about missing symbols. the
reason is that the object files - despite having the same name - contain
different symbols (because each Boost library has its own namespace that
is then encoded into the symbol names). when one of the object files is
overwritten, its symbols are lost and will not be present in the final
libboost.a uberlib.

the fix for this is simply to use library-specific obj folders when libraries
are decomposed.

an example linker error:

Undefined symbols for architecture i386:
  "vtable for boost::filesystem::detail::utf8_codecvt_facet", referenced from:
      boost::filesystem::detail::utf8_codecvt_facet::utf8_codecvt_facet(unsigned long) in AppDelegate.o
  NOTE: a missing vtable usually means the first non-inline virtual member function has no definition.
ld: symbol(s) not found for architecture i386
clang: error: linker command failed with exit code 1 (use -v to see invocation)