fuego-on-ios.xcodeproj now defines the NDEBUG macro
if NDEBUG is not defined and Fuego is run with --quiet, an assertion in SgDebug()
will fail, causing the program to abort. In SgDebug.cpp:

    if (! g_debugStrPtr->good())
    {
        // does not just use a direct SG_ASSERT(g_debugStrPtr->good())
        // in order to allow a breakpoint to be set on the line below.
        SG_ASSERT(false);
    }

The check whether the debug stream is good fails when --quiet is specified.
NDEBUG must be defined so that the SG_ASSERT macro will expand to
nothing.