package fuego-on-ios as an XCFramework (#22)
stitching together a fat binary with lipo from the slices in a simulator
and a device build no longer works, because with Xcode 12 both the
simulator and the device builds now contain an arm64 slice. fat binaries
are limited inasmuch as they can hold only one slice per architecture.
XCFramework is a new packaging format that overcomes this limitation.

why does the simulator build contain an arm64 slice? because Xcode 12
supports Apple Silicon machines, which are based on ARM CPUs. this means
that a simulator on an Apple Silicon machine runs on ARM.