disable CLANG_WARN_QUOTED_INCLUDE_IN_FRAMEWORK_HEADER / -Wquoted-include-in-framework-header
the warning was enabled in the previous commit after upgrading the Fuego
Xcode project to Xcode 26.0.1. the warning is emitted a large number of
times because many Boost header files contain include statements like
this:

  #include "boost/date_time/date.hpp"

correct would be

  #include <boost/date_time/date.hpp>

because this project has no control over the Boost framework headers
(short of patching them to use the correct form), it seems better to
disable the warning so that it does not hide other, possibly more
important warnings.